Invento/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
mysqlclient==2.2.4
dj-rest-auth-6.0.0
djangorestframework-simplejwt-5.3.1 pyjwt-2.9.0